feat: refactored axios endpoints terminology
