+++ src/api/endpoints/job-categories/getCategories.endpoint.ts
