+++ src/api/endpoints/job-categories/getCategory.endpoint.ts
