+++ src/api/endpoints/skills/getSingleSkill.endpoint.ts
